Initial work towards implementing a Flatbush.
This is a static Rtree that should have a low memory footprint
and fast lookup times, and fast construction times.  It stores
the degree k tree of bounding boxes as a single array of
Envelopes, instead of a tree of objects with children.  Traversing
the tree is done by integer index calculations.

It is a static tree, immutable once constructed.  It can be
constructed either without ordering the input items (if their
order is already known to be amenable to construction), or
ordering via Hilbert Curve.  This gives the appropriate
hierarchical clustering properties.