Add flatbush.find_intersection_candidates
Also, removing the reference to items in flatbush.  Flatbush will
only return indices of the items, which the owner can use to reference
the item.  This makes the ownership/lifetime model of the flatbush
significantly simpler, although it would allow the caller to modify
the items vec which would cause the indices to be out of sync.